Update scripts.js
If applied, this commit load some data from an external API and display it to my users. The user should also be able to view details for the data they selected in the console. For this I implemented a modal as a details view for each Pokémon in my list.